// apps/nestjs-backend/src/config/config-key.enum.ts
export enum ConfigKey {
  NODE_ENV = 'NODE_ENV',
  FRONTEND_HOST = 'FRONTEND_HOST',
  PORT = 'PORT',
  ENABLE_SWAGGER = 'ENABLE_SWAGGER',

  DATABASE_URL = 'DATABASE_URL',
  POSTGRES_TIMEZONE = 'POSTGRES_TIMEZONE',
  POSTGRES_DB_NAME = 'POSTGRES_DB_NAME',
  POSTGRES_PASSWORD = 'POSTGRES_PASSWORD',
  POSTGRES_PORT = 'POSTGRES_PORT',
  POSTGRES_HOST = 'POSTGRES_HOST',
  POSTGRES_USER = 'POSTGRES_USER',
  POSTGRES_DEBUG_MODE = 'POSTGRES_DEBUG_MODE',

  REDIS_URL = 'REDIS_URL',
  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',
  REDIS_PASSWORD = 'REDIS_PASSWORD',

  JWT_SECRET = 'JWT_SECRET',
  JWT_EXPIRES_IN = 'JWT_EXPIRES_IN',
}